upgraded ArgoCD version from 1.7.x to 1.8
