+++ pkg/controller/argocd/statefulset_test.go
@@ -5,6 +5,9 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/google/go-cmp/cmp"
+	appsv1 "k8s.io/api/apps/v1"
+
 	"gotest.tools/assert"
 
 	"k8s.io/apimachinery/pkg/types"
@@ -21,7 +24,7 @@ func TestReconcileArgoCD_reconcileRedisStatefulSet_HA_disabled(t *testing.T) {
 
 	a := makeTestArgoCD()
 	r := makeTestReconciler(t, a)
-	s := newStatefulSetWithSuffix("redis-ha-server", "redis", a)
+	s := newStatefulSetWithSuffix("redis-ha", "redis", a)
 
 	assert.NilError(t, r.reconcileRedisStatefulSet(a))
 	// resource Creation should fail as HA was disabled
@@ -33,7 +36,7 @@ func TestReconcileArgoCD_reconcileRedisStatefulSet_HA_enabled(t *testing.T) {
 
 	a := makeTestArgoCD()
 	r := makeTestReconciler(t, a)
-	s := newStatefulSetWithSuffix("redis-ha-server", "redis", a)
+	s := newStatefulSetWithSuffix("redis-ha", "redis", a)
 
 	a.Spec.HA.Enabled = true
 	// test resource is Created when HA is enabled
@@ -51,4 +54,74 @@ func TestReconcileArgoCD_reconcileRedisStatefulSet_HA_enabled(t *testing.T) {
 	a.Spec.HA.Enabled = false
 	assert.NilError(t, r.reconcileRedisStatefulSet(a))
 	assert.ErrorContains(t, r.client.Get(context.TODO(), types.NamespacedName{Name: s.Name, Namespace: a.Namespace}, s), "not found")
+}
+
+func TestReconcileArgoCD_reconcileApplicationController(t *testing.T) {
+	logf.SetLogger(logf.ZapLogger(true))
+	a := makeTestArgoCD()
+	r := makeTestReconciler(t, a)
+
+	assert.NilError(t, r.reconcileApplicationControllerStatefulSet(a))
+
+	ss := &appsv1.StatefulSet{}
+	assert.NilError(t, r.client.Get(
+		context.TODO(),
+		types.NamespacedName{
+			Name:      "argocd-application-controller",
+			Namespace: a.Namespace,
+		},
+		ss))
+	command := ss.Spec.Template.Spec.Containers[0].Command
+	want := []string{
+		"argocd-application-controller",
+		"--operation-processors", "10",
+		"--redis", "argocd-redis.argocd.svc.cluster.local:6379",
+		"--repo-server", "argocd-repo-server.argocd.svc.cluster.local:8081",
+		"--status-processors", "20"}
+	if diff := cmp.Diff(want, command); diff != "" {
+		t.Fatalf("reconciliation failed:\n%s", diff)
+	}
+}
+
+func TestReconcileArgoCD_reconcileApplicationController_withUpdate(t *testing.T) {
+	logf.SetLogger(logf.ZapLogger(true))
+	a := makeTestArgoCD()
+	r := makeTestReconciler(t, a)
+
+	assert.NilError(t, r.reconcileApplicationControllerStatefulSet(a))
+
+	a = makeTestArgoCD(controllerProcessors(30))
+	assert.NilError(t, r.reconcileApplicationControllerStatefulSet(a))
+
+	ss := &appsv1.StatefulSet{}
+	assert.NilError(t, r.client.Get(
+		context.TODO(),
+		types.NamespacedName{
+			Name:      "argocd-application-controller",
+			Namespace: a.Namespace,
+		},
+		ss))
+	command := ss.Spec.Template.Spec.Containers[0].Command
+	want := []string{
+		"argocd-application-controller",
+		"--operation-processors", "10",
+		"--redis", "argocd-redis.argocd.svc.cluster.local:6379",
+		"--repo-server", "argocd-repo-server.argocd.svc.cluster.local:8081",
+		"--status-processors", "30"}
+	if diff := cmp.Diff(want, command); diff != "" {
+		t.Fatalf("reconciliation failed:\n%s", diff)
+	}
+}
+
+func TestReconcileArgoCD_reconcileApplicationController_withUpgrade(t *testing.T) {
+	logf.SetLogger(logf.ZapLogger(true))
+	a := makeTestArgoCD()
+	r := makeTestReconciler(t, a)
+
+	deploy := newDeploymentWithSuffix("application-controller", "application-controller", a)
+	assert.NilError(t, r.client.Create(context.TODO(), deploy))
+
+	assert.NilError(t, r.reconcileApplicationControllerStatefulSet(a))
+	err := r.client.Get(context.TODO(), types.NamespacedName{Name: deploy.Name, Namespace: deploy.Namespace}, deploy)
+	assert.ErrorContains(t, err, "not found")
 }
